fix: Move name parsing code to data_types2
