+++ ioxd_common/Cargo.toml
@@ -11,6 +11,7 @@ edition = "2021"
 # Workspace dependencies, in alphabetical order
 clap_blocks = { path = "../clap_blocks" }
 data_types = { path = "../data_types" }
+data_types2 = { path = "../data_types2" }
 dml = { path = "../dml" }
 metric = { path = "../metric" }
 observability_deps = { path = "../observability_deps" }
