+++ ioxd_common/src/http/dml.rs
@@ -2,19 +2,16 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use chrono::Utc;
-use hyper::{Body, Method, Request, Response, StatusCode};
-use serde::Deserialize;
-use snafu::{OptionExt, ResultExt, Snafu};
-
-use data_types::{
-    names::{org_and_bucket_to_database, OrgBucketMappingError},
-    non_empty::NonEmptyString,
-    DatabaseName,
+use data_types2::{
+    org_and_bucket_to_database, DatabaseName, NonEmptyString, OrgBucketMappingError,
 };
 use dml::{DmlDelete, DmlMeta, DmlOperation, DmlWrite};
+use hyper::{Body, Method, Request, Response, StatusCode};
 use mutable_batch_lp::LinesConverter;
 use observability_deps::tracing::debug;
 use predicate::delete_predicate::{parse_delete_predicate, parse_http_delete_request};
+use serde::Deserialize;
+use snafu::{OptionExt, ResultExt, Snafu};
 
 use crate::{http::utils::parse_body, server_type::ServerType};
